requirements.txt
ipykernel
openai
python-dotenv
streamlit
streamlit_webrtc
pydub
fpdf
imageio-ffmpeg